+++ test/python/test_thirdparty.py
@@ -2,7 +2,7 @@ import logging
 import unittest
 from config import AppConfig
 from zensols.config import ImportConfigFactory
-from zensols.nlp import LanguageResourceFactory
+import zensols.nlp
 
 logger = logging.getLogger(__name__)
 
@@ -10,12 +10,12 @@ logger = logging.getLogger(__name__)
 class TestThirdParty(unittest.TestCase):
     def setUp(self):
         self.config = AppConfig('stemmer')
-        self.fac = LanguageResourceFactory(self.config)
+        self.fac = ImportConfigFactory(self.config)
 
     def test_stemmer(self):
         tnfac = ImportConfigFactory(self.config)
         sent = 'Bobby is fast and runs with dogs, armies, and sheep from the police.'
-        lr = self.fac.instance(token_normalizer=tnfac.instance('nonorm_token_normalizer'))
+        lr = self.fac.instance('default_langres', token_normalizer=tnfac.instance('nonorm_token_normalizer'))
         doc = lr.parse(sent)
         feats = tuple(lr.features(doc))
         self.assertEqual(('Bobby', 'is', 'fast', 'and', 'runs', 'with',
@@ -26,7 +26,7 @@ class TestThirdParty(unittest.TestCase):
                           ',', 'army', ',', 'and', 'sheep', 'from', 'the',
                           'police', '.'),
                          tuple(map(lambda f: f.lemma, feats)))
-        lr = self.fac.instance(token_normalizer=tnfac.instance('stemmer_token_normalizer'))
+        lr = self.fac.instance('default_langres', token_normalizer=tnfac.instance('stemmer_token_normalizer'))
         feats = tuple(lr.features(doc))
         self.assertEqual(('bobbi', 'is', 'fast', 'and', 'run', 'with', 'dog',
                           ',', 'armi', ',', 'and', 'sheep', 'from', 'the',