+++ src/python/zensols/nlp/__init__.py
@@ -1,6 +1,4 @@
 from zensols.nlp.norm import *
 from zensols.nlp.stemmer import *
-from zensols.nlp.word2vec import *
 from zensols.nlp.feature import *
-from zensols.nlp.fnorm import *
 from zensols.nlp.lang import *