+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+### 
+
+
 ## [0.0.8] - 2020-04-14
 ### Changed
 - Upgrade to `spaCy` 2.2.4 and `textacy` 0.10.0