+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Feature to parse whole special tokens.
 
+### Removed
+- Moved word2vec embedding (`word2vec.py`) to [zensols.deepnlp] library.
+- Moved feature normalization (`fnorm.py`) to [zensols.deepnlp] library.
+
 
 ## [0.0.8] - 2020-04-14
 ### Changed
@@ -82,3 +86,4 @@ Data classes are now used so Python 3.7 is now a requirement.
 [0.0.1]: https://github.com/plandes/nlparse/compare/v0.0.0...v0.0.1
 
 [NLTK]: https://www.nltk.org
+[zensols.deepnlp]: https://github.com/plandes/deepnlp