+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [0.0.9] - 2020-05-10
 ### Changed
 - Home/master move lemmatizing out of default token normalizer.
 - Update super method calls to modern (at least) Python 3.7.
@@ -75,7 +76,8 @@ Data classes are now used so Python 3.7 is now a requirement.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/nlparse/compare/v0.0.8...HEAD
+[Unreleased]: https://github.com/plandes/nlparse/compare/v0.0.9...HEAD
+[0.0.9]: https://github.com/plandes/nlparse/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/plandes/nlparse/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/plandes/nlparse/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/plandes/nlparse/compare/v0.0.5...v0.0.6